requirements.txt
accelerate==0.11.0
comet-ml==3.31.12
matplotlib==3.3.0
datasets==2.5.2
tqdm==4.64.1
tensorboard==2.10.1
tensorboardX==2.5.1
torchvision==0.11.2+cu111
tensorflow-datasets==4.6.0
einops==0.4.1
pytorch-fid==0.2.1
joblib==1.1.0
PyYAML==6.0
kaggle==1.5.12
wandb==0.13.4
torchsummary==1.5.1
torchinfo==1.7.1
